requirements.txt
Django==2.1
Markdown==2.6.11
django-mdeditor==0.1.8
django-cors-headers==2.4.0
Pillow==5.3.0
uwsgi==2.0.17.1
Pygments==2.2.0
django-taggit==0.23.0
